try to remove old binary at entry
